+++ src/reduct/storage/entry.h
@@ -23,10 +23,8 @@ class IEntry {
    * Options
    */
   struct Options {
-    std::string name;            // name of entry
-    std::filesystem::path path;  // path to entry (path to bucket)
-    size_t max_block_size;       // max block quota_size after that we create a new one
-    size_t max_block_records;    // max number of records in a block
+    size_t max_block_size;     // max block quota_size after that we create a new one
+    size_t max_block_records;  // max number of records in a block
 
     bool operator<=>(const Options& rhs) const = default;
   };
@@ -88,12 +86,18 @@ class IEntry {
   [[nodiscard]] virtual const Options& GetOptions() const = 0;
 
   /**
+   * @brief Set options
+   * @return
+   */
+  virtual void SetOptions(const Options&) = 0;
+
+  /**
    * @brief Creates a new entry.
    * Directory path/name must be empty
    * @param options
    * @return pointer to entre or nullptr if failed to create
    */
-  static std::unique_ptr<IEntry> Build(Options options);
+  static std::unique_ptr<IEntry> Build(std::string_view name, const std::filesystem::path& path, Options options);
 
   /**
    * @brief Restores entry from dir